Create new package structure for `runtime`
As part of https://github.com/deepmap/oapi-codegen/issues/1142 we can
start to reduce our transitive dependencies by refactoring the
`pkg/runtime` as it exists in deepmap/oapi-codegen.

This allows us to introduce a breaking change (tweaking the structure of
this package, including moving the `strictmiddleware` concepts) by
creating a fresh package,

By moving `runtime` to a separate package, we can allow more frequent
updates to the runtime-specific code, without tying it directly to the
version of the code generator in use.

This also uses package structure to allow reducing the transitive
dependencies introduced by `strictmiddleware.go`, where Go will prune
the tree if it's unused.